requirements.txt
Mezzanine==4.1.0
django-allauth
django-robots
